feat: change dialog for modal bottom sheet
