+++ lib/features/workouts/widgets/widgets.dart
@@ -1,4 +1,4 @@
 export 'add_workout_card.dart';
 export 'workout_card.dart';
-export 'workout_options_dialog.dart';
+export 'workout_options_bottom_sheet.dart';
 export 'workouts_body.dart';
